Refactoring around async and rputs
