+++ src/http/ngx_http_mruby_module.c
@@ -774,37 +774,6 @@ void ngx_http_mruby_read_request_body_cb(ngx_http_request_t *r)
   }
 }
 
-// TODO: Support rputs by multi directive
-ngx_int_t ngx_mrb_finalize_rputs(ngx_http_request_t *r, ngx_http_mruby_ctx_t *ctx)
-{
-  ngx_int_t rc;
-  ngx_mrb_rputs_chain_list_t *chain;
-
-  chain = ctx->rputs_chain;
-  if (chain == NULL) {
-    ngx_log_error(NGX_LOG_INFO, r->connection->log, 0,
-                  "%s INFO %s:%d: mrb_run info: rputs_chain is null and return NGX_OK", MODULE_NAME, __func__,
-                  __LINE__);
-    if (r->headers_out.status >= 100) {
-      rc = r->headers_out.status;
-    } else {
-      rc = NGX_OK;
-    }
-  }
-  if (r->headers_out.status == NGX_HTTP_OK || !(*chain->last)->buf->last_buf) {
-    r->headers_out.status = NGX_HTTP_OK;
-    (*chain->last)->buf->last_buf = 1;
-    ngx_http_send_header(r);
-    ngx_http_output_filter(r, chain->out);
-    ngx_http_set_ctx(r, NULL, ngx_http_mruby_module);
-    rc = NGX_OK;
-  } else {
-    rc = r->headers_out.status;
-  }
-
-  return rc;
-}
-
 ngx_int_t ngx_mrb_run(ngx_http_request_t *r, ngx_mrb_state_t *state, ngx_mrb_code_t *code, ngx_flag_t cached,
                       ngx_str_t *result)
 {
@@ -869,11 +838,7 @@ ngx_int_t ngx_mrb_run(ngx_http_request_t *r, ngx_mrb_state_t *state, ngx_mrb_cod
   }
 
 #ifdef NGX_USE_MRUBY_ASYNC
-  mrb_value *fiber = NULL;
-  fiber = (mrb_value *)ngx_palloc(r->pool, sizeof(mrb_value));
-  ngx_mrb_prepare_fiber(state->mrb, code->proc, fiber);
-
-  if (mrb_test(ngx_mrb_run_fiber(state->mrb, fiber, &mrb_result))) {
+  if (mrb_test(ngx_mrb_start_fiber(r, state->mrb, code->proc, &mrb_result))) {
     return NGX_DONE;
   }
 #else
